pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "devault"
version = "0.5.0"
description = "Developer-focused backup platform — file backup MVP"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "sqlalchemy>=2.0.36",
  "psycopg[binary]>=3.2.0",
  "alembic>=1.14.0",
  "grpcio>=1.80.0",
  "grpcio-health-checking>=1.80.0",
  "protobuf>=5.26.0",
  "redis>=5.2.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.6.0",
  "boto3>=1.35.0",
  "pathspec>=0.12.1",
  "httpx>=0.28.0",
  "typer>=0.15.0",
  "apscheduler>=3.10.0",
  "prometheus-client>=0.21.0",
  "croniter>=2.0.0",
  "python-multipart>=0.0.9",
  "packaging>=24.0",
  "PyJWT[crypto]>=2.9.0",
  "cryptography>=43.0.0",
  "argon2-cffi>=23.1.0",
  "pyotp>=2.9.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=6.0.0",
  "grpcio-tools>=1.80.0",
  "fakeredis>=2.23.0",
]

[project.scripts]
devault = "devault.cli.main:main"
devault-admin = "devault.admin:main"
devault-scheduler = "devault.scheduler.main:main"
devault-agent = "devault.agent.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = ["src/devault/grpc_gen"]
